fix(playerDropped): Multiple Financed Cars Error
This should resolve the error for playTIme being nil when a player leave the server. The error only happens if they have more than one financed vehicles and it is because after the first vehicle gets updated it sets the table entry to an empty table and no longer the join server time.

Also I changed it from setting as an empty table to nil to remove the entry from the table entirely.