Fix Finance menu from showing all vehicles
Finance menu was showing all owned vehicles since balance is 0 by default and not nil in the DB. To fix this I added a small check to the event 'qb-vehicleshop:client:getVehicles' to ensure it did not retrieve vehicles with a equal to zero. 

Now the menu only displays vehicles with a balance rather than every owned vehicle.

Tested and working on my latest qb-core server.